#40 finish implementation of terraform provider

+++ terraform-provider-terratowns/main.go
@@ -207,7 +207,7 @@ func resourceHouseRead(ctx context.Context, d *schema.ResourceData, m interface{
 		d.Set("name", responseData["name"].(string))
 		d.Set("description", responseData["description"].(string))
 		d.Set("domain_name", responseData["domain_name"].(string))
-		d.Set("content_version", responseData["content_version"].(int64))
+		d.Set("content_version", responseData["content_version"].(float64))
 	} else if resp.StatusCode != http.StatusOK {
 		d.SetId("")
 	} else if resp.StatusCode != http.StatusOK {
